Code clean: moved useFetchCategoryByUuid into webapp/service/hooks/categories
