+++ webapp/service/hooks/categories/index.js
@@ -1,0 +1,1 @@
+export { useFetchCategoryByUuid } from './useFetchCategoryByUuid'
