+++ webapp/service/hooks/categories/useFetchCategoryByUuid.js
